+++ src/lib/compiler/scriptBuilder.ts
@@ -3764,6 +3764,21 @@ class ScriptBuilder {
     this._addNL();
   };
 
+  ifDeviceGBA = (truePath = [], falsePath = []) => {
+    const falseLabel = this.getNextLabel();
+    const endLabel = this.getNextLabel();
+    this._addComment(`If Device GBA`);
+    this._stackPushConst(0);
+    this._getMemUInt8(".ARG0", "_is_GBA");
+    this._ifConst(".NE", ".ARG0", 1, falseLabel, 1);
+    this._compilePath(truePath);
+    this._jump(endLabel);
+    this._label(falseLabel);
+    this._compilePath(falsePath);
+    this._label(endLabel);
+    this._addNL();
+  };
+
   ifActorAtPosition = (
     x: number,
     y: number,