+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added event to determine if device is running on SGB
+- Added event to determine if device is a GBA
 
 ### Changed